+++ examples/FollowStrawberry/Program.cs
@@ -7,7 +7,7 @@ namespace FollowStrawberry
 {
 	class Program
 	{
-		static int Main(string[] args)
+		static int Main()
 		{
 			DF.Root.Add(
 				Sprite("strawberry", "strawberry.png")
